Update to allow recommended ref values
Assigning strings to a ref is considered deprecated by the React team.
Instead, it is recommended that you use anonymous fat arrow functions in
refs to assign the reffed component to a variable/property of the
component whose file you're in.

The linter currently errors with a jsx-no-bind error when you do that,
though, so this commit makes an exception for fat arrow functions in
refs while otherwise preserving the linter's behavior (which is
generally what we want).